requirements.txt
streamlit
openai
Pillow
opencv-python
python-dotenv
langchain
langchain-openai
